+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to the taptempo project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [TapTempo Fortran dev]
+
+### Added
+- A new option:
+  -o, --output          save the results in the taptempo.txt file
+
+### Fixed
+- More robust command line options.
 
 ## [TapTempo Fortran 0.9] 2021-10-28
 Initial commit.